Reduce platform stack ingress blast radius
Decompose the monolithic PlatformStack ingress surface into focused L3 constructs for SPA, REST API, WAF, and AgentCore gateway concerns while preserving synthesized resource shape and existing test expectations. Follow-up cleanup also removes the deprecated DynamoDB PITR flag usage so the CDK verification path is warning-free again.

Constraint: Existing CloudFormation outputs and security-sensitive synth shape must remain stable during the extraction

Rejected: Leave output logical IDs nested under child constructs | broke existing synth assertions and increased template drift risk

Rejected: Defer DynamoDB PITR deprecation cleanup | would leave warning noise and known debt in the verification path

Confidence: high

Scope-risk: moderate

Reversibility: clean

Directive: Keep PlatformStack as the composition root for ingress concerns; add new ingress resources in focused constructs unless a cross-cutting dependency forces reintegration

Tested: npm run build

Tested: npx jest --runInBand --runTestsByPath test/platform-ingress-constructs.test.ts

Tested: npx jest --runInBand --runTestsByPath test/platform-stack.test.ts
