+++ infra/cdk/lib/platform-stack.ts
@@ -14,7 +14,6 @@
  */
 import * as cdk from 'aws-cdk-lib';
 import * as apigateway from 'aws-cdk-lib/aws-apigateway';
-import * as acm from 'aws-cdk-lib/aws-certificatemanager';
 import * as cloudfront from 'aws-cdk-lib/aws-cloudfront';
 import * as cloudwatch from 'aws-cdk-lib/aws-cloudwatch';
 import * as codedeploy from 'aws-cdk-lib/aws-codedeploy';
@@ -23,12 +22,9 @@ import * as ec2 from 'aws-cdk-lib/aws-ec2';
 import * as iam from 'aws-cdk-lib/aws-iam';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 import * as s3 from 'aws-cdk-lib/aws-s3';
-import * as logs from 'aws-cdk-lib/aws-logs';
 import * as secretsmanager from 'aws-cdk-lib/aws-secretsmanager';
 import * as sqs from 'aws-cdk-lib/aws-sqs';
 import * as ssm from 'aws-cdk-lib/aws-ssm';
-import * as sfn from 'aws-cdk-lib/aws-stepfunctions';
-import * as sfn_tasks from 'aws-cdk-lib/aws-stepfunctions-tasks';
 import * as wafv2 from 'aws-cdk-lib/aws-wafv2';
 import * as s3assets from 'aws-cdk-lib/aws-s3-assets';
 import { Template } from 'aws-cdk-lib/assertions';
@@ -38,6 +34,10 @@ import * as path from 'path';
 import { createPlatformCompute } from './platform-compute';
 import { resolveEntraConfiguration } from './entra-config';
 import { createPlatformStorage } from './platform-storage';
+import { PlatformApi } from './platform-api';
+import { PlatformGateway } from './platform-gateway';
+import { PlatformSpa } from './platform-spa';
+import { PlatformWaf } from './platform-waf';
 import { TenantStack } from './tenant-stack';
 
 const TENANT_AUTHORIZED_RUNTIME_REGIONS = ['eu-west-1', 'eu-central-1'] as const;
@@ -278,577 +278,27 @@ export class PlatformStack extends cdk.Stack {
       provisionedConcurrentExecutions: 10,
     });
 
-    const restAuthorizer = new apigateway.TokenAuthorizer(this, 'RestTokenAuthorizer', {
-      handler: authoriserAlias,
-      identitySource: apigateway.IdentitySource.header('Authorization'),
-      resultsCacheTtl: cdk.Duration.minutes(5),
+    const platformSpa = new PlatformSpa(this, 'PlatformSpa', {
+      envName: env,
+      spaDomainName,
+      spaCertificateArn,
     });
+    this.spaDistribution = platformSpa.spaDistribution;
 
-    const spaBucket = new s3.Bucket(this, 'SpaBucket', {
-      blockPublicAccess: s3.BlockPublicAccess.BLOCK_ALL,
-      encryption: s3.BucketEncryption.S3_MANAGED,
-      enforceSSL: true,
-      versioned: true,
+    const platformApi = new PlatformApi(this, 'PlatformApi', {
+      envName: env,
+      spaAllowedOrigin: platformSpa.spaAllowedOrigin,
+      apiDomainName,
+      apiCertificateArn,
+      authoriserAlias,
+      tenantMgmtFn: this.tenantMgmtFn,
+      webhookRegistryFn: this.webhookRegistryFn,
+      agentRegistryFn: this.agentRegistryFn,
+      adminOpsFn: this.adminOpsFn,
+      bridgeAlias,
+      bffFn: this.bffFn,
     });
-
-    const spaLogBucket = new s3.Bucket(this, 'SpaLogBucket', {
-      bucketName: `platform-spa-logs-${env}`,
-      blockPublicAccess: s3.BlockPublicAccess.BLOCK_ALL,
-      encryption: s3.BucketEncryption.S3_MANAGED,
-      enforceSSL: true,
-      removalPolicy: cdk.RemovalPolicy.RETAIN,
-      objectOwnership: s3.ObjectOwnership.BUCKET_OWNER_PREFERRED,
-      accessControl: s3.BucketAccessControl.LOG_DELIVERY_WRITE,
-    });
-
-    const spaResponseHeadersPolicy = new cloudfront.CfnResponseHeadersPolicy(
-      this,
-      'SpaCspResponseHeadersPolicy',
-      {
-        responseHeadersPolicyConfig: {
-          name: `${this.stackName}-spa-security-headers`,
-          comment: 'Security headers for platform SPA',
-          securityHeadersConfig: {
-            contentSecurityPolicy: {
-              contentSecurityPolicy:
-                "default-src 'self'; object-src 'none'; frame-ancestors 'none'; base-uri 'self'; connect-src 'self' https:; img-src 'self' data: https:; style-src 'self' 'unsafe-inline'; script-src 'self';",
-              override: true,
-            },
-            frameOptions: {
-              frameOption: 'DENY',
-              override: true,
-            },
-            strictTransportSecurity: {
-              accessControlMaxAgeSec: 31536000,
-              includeSubdomains: true,
-              preload: true,
-              override: true,
-            },
-            contentTypeOptions: {
-              override: true,
-            },
-            referrerPolicy: {
-              referrerPolicy: 'same-origin',
-              override: true,
-            },
-            xssProtection: {
-              protection: true,
-              modeBlock: true,
-              override: true,
-            },
-          },
-        },
-      },
-    );
-
-    const spaOriginAccessControl = new cloudfront.CfnOriginAccessControl(this, 'SpaOriginAccessControl', {
-      originAccessControlConfig: {
-        name: `${this.stackName}-spa-oac`,
-        description: 'OAC for SPA bucket origin',
-        originAccessControlOriginType: 's3',
-        signingBehavior: 'always',
-        signingProtocol: 'sigv4',
-      },
-    });
-
-    const spaRouteRewriteFunction = new cloudfront.Function(this, 'SpaRouteRewriteFunction', {
-      comment: 'Rewrite SPA deep links to index.html without masking missing asset failures',
-      code: cloudfront.FunctionCode.fromInline(`
-function handler(event) {
-  var request = event.request;
-  var uri = request.uri || '/';
-  var lastSegment = uri.substring(uri.lastIndexOf('/') + 1);
-
-  if (uri === '/' || lastSegment.indexOf('.') === -1) {
-    request.uri = '/index.html';
-  }
-
-  return request;
-}
-      `),
-    });
-
-    this.spaDistribution = new cloudfront.CfnDistribution(this, 'SpaDistribution', {
-      distributionConfig: {
-        enabled: true,
-        comment: 'Platform SPA distribution',
-        defaultRootObject: 'index.html',
-        httpVersion: 'http2',
-        priceClass: 'PriceClass_100',
-        ipv6Enabled: true,
-        logging: {
-          bucket: spaLogBucket.bucketRegionalDomainName,
-          includeCookies: false,
-          prefix: 'spa-cloudfront/',
-        },
-        origins: [
-          {
-            id: 'SpaS3Origin',
-            domainName: spaBucket.bucketRegionalDomainName,
-            originAccessControlId: spaOriginAccessControl.attrId,
-            s3OriginConfig: {
-              originAccessIdentity: '',
-            },
-          },
-        ],
-        defaultCacheBehavior: {
-          targetOriginId: 'SpaS3Origin',
-          viewerProtocolPolicy: 'redirect-to-https',
-          compress: true,
-          allowedMethods: ['GET', 'HEAD', 'OPTIONS'],
-          cachedMethods: ['GET', 'HEAD', 'OPTIONS'],
-          cachePolicyId: cloudfront.CachePolicy.CACHING_DISABLED.cachePolicyId,
-          responseHeadersPolicyId: spaResponseHeadersPolicy.attrId,
-          functionAssociations: [
-            {
-              eventType: 'viewer-request',
-              functionArn: spaRouteRewriteFunction.functionArn,
-            },
-          ],
-        },
-        cacheBehaviors: [
-          {
-            pathPattern: 'assets/*',
-            targetOriginId: 'SpaS3Origin',
-            viewerProtocolPolicy: 'redirect-to-https',
-            compress: true,
-            allowedMethods: ['GET', 'HEAD', 'OPTIONS'],
-            cachedMethods: ['GET', 'HEAD', 'OPTIONS'],
-            cachePolicyId: cloudfront.CachePolicy.CACHING_OPTIMIZED.cachePolicyId,
-            responseHeadersPolicyId: spaResponseHeadersPolicy.attrId,
-          },
-        ],
-        restrictions: {
-          geoRestriction: {
-            restrictionType: 'none',
-          },
-        },
-        ...(spaDomainName && spaCertificateArn
-          ? {
-              aliases: [spaDomainName],
-              viewerCertificate: {
-                acmCertificateArn: spaCertificateArn,
-                minimumProtocolVersion: 'TLSv1.2_2021',
-                sslSupportMethod: 'sni-only',
-              },
-            }
-          : {
-              viewerCertificate: {
-                cloudFrontDefaultCertificate: true,
-                minimumProtocolVersion: 'TLSv1.2_2021',
-              },
-            }),
-        // No WebACLId is wired here by design. A CloudFront-scope WAF must be managed
-        // via a dedicated global/us-east-1 path, which this repository has not yet
-        // approved under the current ADR-009 region topology.
-      },
-    });
-
-    spaBucket.addToResourcePolicy(
-      new iam.PolicyStatement({
-        sid: 'AllowCloudFrontOacRead',
-        effect: iam.Effect.ALLOW,
-        principals: [new iam.ServicePrincipal('cloudfront.amazonaws.com')],
-        actions: ['s3:GetObject'],
-        resources: [spaBucket.arnForObjects('*')],
-        conditions: {
-          StringEquals: {
-            'AWS:SourceArn': cdk.Fn.join('', [
-              'arn:',
-              cdk.Aws.PARTITION,
-              ':cloudfront::',
-              cdk.Aws.ACCOUNT_ID,
-              ':distribution/',
-              this.spaDistribution.ref,
-            ]),
-          },
-        },
-      }),
-    );
-
-    new ssm.StringParameter(this, 'SpaBucketNameParam', {
-      parameterName: `/platform/spa/${env}/bucket-name`,
-      stringValue: spaBucket.bucketName,
-      description: 'S3 bucket name for the platform SPA',
-    });
-
-    new ssm.StringParameter(this, 'SpaDistributionIdParam', {
-      parameterName: `/platform/spa/${env}/distribution-id`,
-      stringValue: this.spaDistribution.ref,
-      description: 'CloudFront distribution ID for the platform SPA',
-    });
-
-    new cdk.CfnOutput(this, 'SpaBucketName', {
-      value: spaBucket.bucketName,
-      description: 'S3 bucket name for the platform SPA',
-    });
-
-    new cdk.CfnOutput(this, 'SpaDistributionId', {
-      value: this.spaDistribution.ref,
-      description: 'CloudFront distribution ID for the platform SPA',
-    });
-
-    if (spaDomainName) {
-      new ssm.StringParameter(this, 'SpaDomainNameParam', {
-        parameterName: `/platform/spa/${env}/domain-name`,
-        stringValue: spaDomainName,
-        description: 'Custom domain name for the platform SPA CloudFront distribution',
-      });
-
-      new cdk.CfnOutput(this, 'SpaDomainName', {
-        value: spaDomainName,
-        description: 'Custom domain name for the platform SPA',
-      });
-    }
-
-    const spaAllowedOrigin = spaDomainName
-      ? `https://${spaDomainName}`
-      : cdk.Fn.join('', ['https://', this.spaDistribution.attrDomainName]);
-
-    // API Access Log Group (TASK-165)
-    const apiAccessLogGroup = new logs.LogGroup(this, 'ApiAccessLogGroup', {
-      logGroupName: `/aws/apigateway/${this.stackName}-rest-api-access-logs`,
-      retention: logs.RetentionDays.ONE_MONTH,
-      removalPolicy: cdk.RemovalPolicy.DESTROY,
-    });
-
-    apiAccessLogGroup.addMetricFilter('TenantRequestCountFilter', {
-      metricName: 'RequestCount',
-      metricNamespace: 'Platform/API',
-      metricValue: '1',
-      filterPattern: logs.FilterPattern.exists('$.tenantId'),
-      dimensions: {
-        TenantId: '$.tenantId',
-      },
-    });
-
-    apiAccessLogGroup.addMetricFilter('TenantErrorCountFilter', {
-      metricName: 'ErrorCount',
-      metricNamespace: 'Platform/API',
-      metricValue: '1',
-      filterPattern: logs.FilterPattern.numberValue('$.status', '>=', 400),
-      dimensions: {
-        TenantId: '$.tenantId',
-      },
-    });
-
-    apiAccessLogGroup.addMetricFilter('TenantLatencyFilter', {
-      metricName: 'Latency',
-      metricNamespace: 'Platform/API',
-      metricValue: '$.latency',
-      filterPattern: logs.FilterPattern.exists('$.latency'),
-      dimensions: {
-        TenantId: '$.tenantId',
-      },
-    });
-
-    this.api = new apigateway.RestApi(this, 'PlatformRestApi', {
-      restApiName: `${this.stackName}-rest-api`,
-      description: 'Platform northbound REST API (ADR-003)',
-      apiKeySourceType: apigateway.ApiKeySourceType.AUTHORIZER,
-      cloudWatchRole: true,
-      endpointConfiguration: {
-        types: [apigateway.EndpointType.REGIONAL],
-      },
-      defaultCorsPreflightOptions: {
-        allowOrigins: [spaAllowedOrigin],
-        allowMethods: apigateway.Cors.ALL_METHODS,
-        allowHeaders: [
-          'Authorization',
-          'Content-Type',
-          'X-Api-Key',
-          'X-Amz-Date',
-          'X-Amz-Security-Token',
-          'X-Amz-User-Agent',
-        ],
-      },
-      deployOptions: {
-        stageName: 'prod',
-        tracingEnabled: true,
-        metricsEnabled: true,
-        throttlingRateLimit: 100,
-        throttlingBurstLimit: 200,
-        accessLogDestination: new apigateway.LogGroupLogDestination(apiAccessLogGroup),
-        accessLogFormat: apigateway.AccessLogFormat.custom(JSON.stringify({
-          requestId: '$context.requestId',
-          extendedRequestId: '$context.extendedRequestId',
-          ip: '$context.identity.sourceIp',
-          caller: '$context.identity.caller',
-          user: '$context.identity.user',
-          requestTime: '$context.requestTime',
-          httpMethod: '$context.httpMethod',
-          resourcePath: '$context.resourcePath',
-          status: 0, // Placeholder for numeric type in JSON.stringify
-          protocol: '$context.protocol',
-          responseLength: 0, // Placeholder
-          tenantId: '$context.authorizer.tenantid',
-          appId: '$context.authorizer.appid',
-          sub: '$context.authorizer.sub',
-          tier: '$context.authorizer.tier',
-          latency: 0, // Placeholder
-        }).replace(':0', ':$context.status').replace(':0', ':$context.responseLength').replace(':0', ':$context.responseLatency')),
-        dataTraceEnabled: false,
-        loggingLevel: apigateway.MethodLoggingLevel.INFO,
-        methodOptions: {
-          '/v1/agents/{agentName}/invoke/POST': {
-            throttlingRateLimit: 50,
-            throttlingBurstLimit: 100,
-            metricsEnabled: true,
-          },
-          '/v1/jobs/{jobId}/GET': {
-            throttlingRateLimit: 100,
-            throttlingBurstLimit: 200,
-            metricsEnabled: true,
-          },
-          '/v1/bff/token-refresh/POST': {
-            throttlingRateLimit: 30,
-            throttlingBurstLimit: 60,
-            metricsEnabled: true,
-          },
-          '/v1/bff/session-keepalive/POST': {
-            throttlingRateLimit: 120,
-            throttlingBurstLimit: 240,
-            metricsEnabled: true,
-          },
-        },
-      },
-    });
-
-    // Add Gateway Responses for CORS and custom errors (TASK-165)
-    // We use quotes for the header values as they are passed to CloudFormation
-    this.api.addGatewayResponse('Default4xxResponse', {
-      type: apigateway.ResponseType.DEFAULT_4XX,
-      responseHeaders: {
-        'gatewayresponses.header.Access-Control-Allow-Origin': `'${spaAllowedOrigin}'`,
-        'gatewayresponses.header.Access-Control-Allow-Credentials': "'true'",
-        'gatewayresponses.header.Access-Control-Allow-Headers': "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
-        'gatewayresponses.header.Access-Control-Allow-Methods': "'GET,POST,OPTIONS'",
-      },
-      templates: {
-        'application/json': '{"message":$context.error.messageString,"requestId":"$context.requestId"}',
-      },
-    });
-
-    this.api.addGatewayResponse('Default5xxResponse', {
-      type: apigateway.ResponseType.DEFAULT_5XX,
-      responseHeaders: {
-        'gatewayresponses.header.Access-Control-Allow-Origin': `'${spaAllowedOrigin}'`,
-        'gatewayresponses.header.Access-Control-Allow-Credentials': "'true'",
-        'gatewayresponses.header.Access-Control-Allow-Headers': "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
-        'gatewayresponses.header.Access-Control-Allow-Methods': "'GET,POST,OPTIONS'",
-      },
-      templates: {
-        'application/json': '{"message":"Internal server error","requestId":"$context.requestId"}',
-      },
-    });
-
-    this.api.addGatewayResponse('UnauthorizedResponse', {
-      type: apigateway.ResponseType.UNAUTHORIZED,
-      responseHeaders: {
-        'gatewayresponses.header.Access-Control-Allow-Origin': `'${spaAllowedOrigin}'`,
-        'gatewayresponses.header.Access-Control-Allow-Credentials': "'true'",
-      },
-      templates: {
-        'application/json': '{"message":"Unauthorized","requestId":"$context.requestId"}',
-      },
-    });
-
-    this.api.addGatewayResponse('AccessDeniedResponse', {
-      type: apigateway.ResponseType.ACCESS_DENIED,
-      responseHeaders: {
-        'gatewayresponses.header.Access-Control-Allow-Origin': `'${spaAllowedOrigin}'`,
-        'gatewayresponses.header.Access-Control-Allow-Credentials': "'true'",
-      },
-      templates: {
-        'application/json': '{"message":"Access denied","requestId":"$context.requestId"}',
-      },
-    });
-
-    const v1 = this.api.root.addResource('v1');
-    const health = v1.addResource('health');
-    const sessions = v1.addResource('sessions');
-    const agents = v1.addResource('agents');
-    const agentByName = agents.addResource('{agentName}');
-    const agentInvoke = agentByName.addResource('invoke');
-    const jobs = v1.addResource('jobs');
-    const jobById = jobs.addResource('{jobId}');
-    const webhooks = v1.addResource('webhooks');
-    const webhookById = webhooks.addResource('{webhookId}');
-    const bff = v1.addResource('bff');
-    const tokenRefresh = bff.addResource('token-refresh');
-    const sessionKeepalive = bff.addResource('session-keepalive');
-
-    const tenants = v1.addResource('tenants');
-    const tenantById = tenants.addResource('{tenantId}');
-    const auditExport = tenantById.addResource('audit-export');
-    const tenantApiKey = tenantById.addResource('api-key');
-    const tenantApiKeyRotate = tenantApiKey.addResource('rotate');
-    const tenantUsers = tenantById.addResource('users');
-    const tenantUsersInvite = tenantUsers.addResource('invite');
-    const tenantUsersInvites = tenantUsers.addResource('invites');
-    const platform = v1.addResource('platform');
-    const failover = platform.addResource('failover');
-    const quota = platform.addResource('quota');
-    const splitAccounts = quota.addResource('split-accounts');
-    const serviceHealth = platform.addResource('service-health');
-    const billing = platform.addResource('billing');
-    const billingStatus = billing.addResource('status');
-    const ops = platform.addResource('ops');
-
-    // Ops sub-resources
-    const opsTopTenants = ops.addResource('top-tenants');
-    const opsSecurityEvents = ops.addResource('security-events');
-    const opsErrorRate = ops.addResource('error-rate');
-    const opsSecurity = ops.addResource('security');
-    const opsSecurityPage = opsSecurity.addResource('page');
-
-    const opsDlq = ops.addResource('dlq');
-    const opsDlqByName = opsDlq.addResource('{proxy+}');
-
-    const opsTenants = ops.addResource('tenants');
-    const opsTenantById = opsTenants.addResource('{proxy+}');
-
-    const opsJobs = ops.addResource('jobs');
-    const opsJobById = opsJobs.addResource('{proxy+}');
-
-    const securedMethodOptions: apigateway.MethodOptions = {
-      authorizer: restAuthorizer,
-      authorizationType: apigateway.AuthorizationType.CUSTOM,
-      apiKeyRequired: true,
-    };
-
-    const tenantMgmtIntegration = new apigateway.LambdaIntegration(this.tenantMgmtFn, { proxy: true });
-    const webhookRegistryIntegration = new apigateway.LambdaIntegration(this.webhookRegistryFn, {
-      proxy: true,
-    });
-    const agentRegistryIntegration = new apigateway.LambdaIntegration(this.agentRegistryFn, {
-      proxy: true,
-    });
-    const adminOpsIntegration = new apigateway.LambdaIntegration(this.adminOpsFn, { proxy: true });
-    const bridgeIntegration = new apigateway.LambdaIntegration(bridgeAlias, { proxy: true });
-    const bridgeStreamingIntegration = new apigateway.LambdaIntegration(bridgeAlias, {
-      proxy: true,
-      responseTransferMode: apigateway.ResponseTransferMode.STREAM,
-    });
-
-    health.addMethod('GET', tenantMgmtIntegration, securedMethodOptions);
-    sessions.addMethod('GET', tenantMgmtIntegration, securedMethodOptions);
-    tenants.addMethod('POST', tenantMgmtIntegration, securedMethodOptions);
-    tenants.addMethod('GET', tenantMgmtIntegration, securedMethodOptions);
-
-    tenantById.addMethod('GET', tenantMgmtIntegration, securedMethodOptions);
-    tenantById.addMethod('PATCH', tenantMgmtIntegration, securedMethodOptions);
-    tenantById.addMethod('DELETE', tenantMgmtIntegration, securedMethodOptions);
-
-    auditExport.addMethod('GET', tenantMgmtIntegration, securedMethodOptions);
-    tenantApiKeyRotate.addMethod('POST', tenantMgmtIntegration, securedMethodOptions);
-    tenantUsersInvite.addMethod('POST', tenantMgmtIntegration, securedMethodOptions);
-    tenantUsersInvites.addMethod('GET', tenantMgmtIntegration, securedMethodOptions);
-
-    failover.addMethod('POST', adminOpsIntegration, securedMethodOptions);
-    quota.addMethod('GET', adminOpsIntegration, securedMethodOptions);
-    splitAccounts.addMethod('POST', adminOpsIntegration, securedMethodOptions);
-    serviceHealth.addMethod('GET', adminOpsIntegration, securedMethodOptions);
-    billingStatus.addMethod('GET', adminOpsIntegration, securedMethodOptions);
-
-    // Wire all ops routes
-    opsTopTenants.addMethod('GET', adminOpsIntegration, securedMethodOptions);
-    opsSecurityEvents.addMethod('GET', adminOpsIntegration, securedMethodOptions);
-    opsErrorRate.addMethod('GET', adminOpsIntegration, securedMethodOptions);
-    opsSecurityPage.addMethod('POST', adminOpsIntegration, securedMethodOptions);
-    opsDlqByName.addMethod('ANY', adminOpsIntegration, securedMethodOptions);
-    opsTenantById.addMethod('ANY', adminOpsIntegration, securedMethodOptions);
-    opsJobById.addMethod('ANY', adminOpsIntegration, securedMethodOptions);
-
-    webhooks.addMethod('GET', webhookRegistryIntegration, securedMethodOptions);
-    webhooks.addMethod('POST', webhookRegistryIntegration, securedMethodOptions);
-    webhookById.addMethod('DELETE', webhookRegistryIntegration, securedMethodOptions);
-
-    const platformAgents = platform.addResource('agents');
-    const platformAgentByName = platformAgents.addResource('{agentName}');
-    const platformAgentVersions = platformAgentByName.addResource('versions');
-    const platformAgentVersion = platformAgentVersions.addResource('{version}');
-
-    platformAgents.addMethod('GET', agentRegistryIntegration, securedMethodOptions);
-    platformAgents.addMethod('POST', agentRegistryIntegration, securedMethodOptions);
-    platformAgentVersion.addMethod('PATCH', agentRegistryIntegration, securedMethodOptions);
-
-    agents.addMethod('GET', bridgeIntegration, securedMethodOptions);
-    agentByName.addMethod('GET', bridgeIntegration, securedMethodOptions);
-    agentInvoke.addMethod('POST', bridgeStreamingIntegration, securedMethodOptions);
-    jobById.addMethod(
-      'GET',
-      bridgeIntegration,
-      securedMethodOptions,
-    );
-    tokenRefresh.addMethod(
-      'POST',
-      new apigateway.LambdaIntegration(this.bffFn, { proxy: true }),
-      securedMethodOptions,
-    );
-    sessionKeepalive.addMethod(
-      'POST',
-      new apigateway.LambdaIntegration(this.bffFn, { proxy: true }),
-      securedMethodOptions,
-    );
-
-    const usagePlanDefinitions = [
-      {
-        id: 'BasicUsagePlan',
-        name: 'basic',
-        rateLimit: 10,
-        burstLimit: 100,
-        quotaLimit: 1000,
-      },
-      {
-        id: 'StandardUsagePlan',
-        name: 'standard',
-        rateLimit: 50,
-        burstLimit: 500,
-        quotaLimit: 10_000,
-      },
-      {
-        id: 'PremiumUsagePlan',
-        name: 'premium',
-        rateLimit: 500,
-        burstLimit: 2_000,
-      },
-    ];
-
-    new ssm.StringParameter(this, 'RestApiIdParam', {
-      parameterName: `/platform/core/${env}/rest-api-id`,
-      stringValue: this.api.restApiId,
-      description: 'REST API ID for the platform northbound API',
-    });
-
-    for (const plan of usagePlanDefinitions) {
-      const usagePlan = new apigateway.UsagePlan(this, plan.id, {
-        name: `${this.stackName}-${plan.name}`,
-        throttle: {
-          rateLimit: plan.rateLimit,
-          burstLimit: plan.burstLimit,
-        },
-        quota:
-          plan.quotaLimit === undefined
-            ? undefined
-            : {
-                limit: plan.quotaLimit,
-                period: apigateway.Period.DAY,
-              },
-        apiStages: [
-          {
-            api: this.api,
-            stage: this.api.deploymentStage,
-          },
-        ],
-      });
-
-      new ssm.StringParameter(this, `${plan.id}IdParam`, {
-        parameterName: `/platform/core/${env}/usage-plan-${plan.name}-id`,
-        stringValue: usagePlan.usagePlanId,
-        description: `Usage plan ID for ${plan.name} tier`,
-      });
-    }
+    this.api = platformApi.api;
 
     new ssm.StringParameter(this, 'BridgeLambdaRoleArnParam', {
       parameterName: `/platform/core/${env}/bridge-lambda-role-arn`,
@@ -856,240 +306,21 @@ function handler(event) {
       description: 'IAM role ARN for the Bridge Lambda function',
     });
 
-    this.apiWebAcl = new wafv2.CfnWebACL(this, 'ApiWebAcl', {
-      name: `${this.stackName}-api-waf`,
-      defaultAction: { allow: {} },
-      scope: 'REGIONAL',
-      visibilityConfig: {
-        cloudWatchMetricsEnabled: true,
-        metricName: `${this.stackName}-api-waf`,
-        sampledRequestsEnabled: true,
-      },
-      rules: [
-        {
-          name: 'AWSManagedRulesCommonRuleSet',
-          priority: 0,
-          overrideAction: { none: {} },
-          statement: {
-            managedRuleGroupStatement: {
-              vendorName: 'AWS',
-              name: 'AWSManagedRulesCommonRuleSet',
-            },
-          },
-          visibilityConfig: {
-            cloudWatchMetricsEnabled: true,
-            metricName: 'aws-managed-common',
-            sampledRequestsEnabled: true,
-          },
-        },
-        {
-          name: 'UkIpRateLimit',
-          priority: 1,
-          action: { block: {} },
-          statement: {
-            rateBasedStatement: {
-              aggregateKeyType: 'IP',
-              limit: 2000,
-              scopeDownStatement: {
-                geoMatchStatement: {
-                  countryCodes: ['GB'],
-                },
-              },
-            },
-          },
-          visibilityConfig: {
-            cloudWatchMetricsEnabled: true,
-            metricName: 'uk-ip-rate-limit',
-            sampledRequestsEnabled: true,
-          },
-        },
-        {
-          name: 'BlockSqlmapUserAgent',
-          priority: 2,
-          action: { block: {} },
-          statement: {
-            byteMatchStatement: {
-              fieldToMatch: {
-                singleHeader: {
-                  Name: 'user-agent',
-                },
-              },
-              positionalConstraint: 'CONTAINS',
-              searchString: 'sqlmap',
-              textTransformations: [
-                {
-                  priority: 0,
-                  type: 'LOWERCASE',
-                },
-              ],
-            },
-          },
-          visibilityConfig: {
-            cloudWatchMetricsEnabled: true,
-            metricName: 'block-sqlmap-user-agent',
-            sampledRequestsEnabled: true,
-          },
-        },
-      ],
+    const platformWaf = new PlatformWaf(this, 'PlatformWaf', {
+      api: this.api,
     });
+    this.apiWebAcl = platformWaf.apiWebAcl;
 
-    new wafv2.CfnWebACLAssociation(this, 'ApiWebAclAssociation', {
-      resourceArn: this.api.deploymentStage.stageArn,
-      webAclArn: this.apiWebAcl.attrArn,
+    const platformGateway = new PlatformGateway(this, 'PlatformGateway', {
+      enforcementMode: gatewayPolicyConfiguration.enforcementMode,
+      policyEngineName: gatewayPolicyConfiguration.policyEngineName,
+      policyName: gatewayPolicyConfiguration.policyName,
+      requestInterceptorFn: this.requestInterceptorFn,
+      responseInterceptorFn: this.responseInterceptorFn,
     });
-
-    // --- API Gateway custom domain (issue #164) ---
-    // When apiDomainName + apiCertificateArn are provided, a regional custom domain
-    // is added to the REST API. The ACM certificate must be in the same region
-    // (eu-west-2) for regional API Gateway endpoints. The caller is responsible for
-    // creating a CNAME or alias DNS record pointing apiDomainName to the regional
-    // domain name output.
-    if (apiDomainName && apiCertificateArn) {
-      const apiCustomDomain = new apigateway.DomainName(this, 'ApiCustomDomain', {
-        domainName: apiDomainName,
-        certificate: acm.Certificate.fromCertificateArn(this, 'ApiCertificate', apiCertificateArn),
-        endpointType: apigateway.EndpointType.REGIONAL,
-        securityPolicy: apigateway.SecurityPolicy.TLS_1_2,
-      });
-
-      new apigateway.BasePathMapping(this, 'ApiBasePathMapping', {
-        domainName: apiCustomDomain,
-        restApi: this.api,
-      });
-
-      new ssm.StringParameter(this, 'ApiDomainNameParam', {
-        parameterName: `/platform/core/${env}/api-domain-name`,
-        stringValue: apiDomainName,
-        description: 'Custom domain name for the platform REST API',
-      });
-
-      new ssm.StringParameter(this, 'ApiRegionalDomainNameParam', {
-        parameterName: `/platform/core/${env}/api-regional-domain-name`,
-        stringValue: apiCustomDomain.domainNameAliasDomainName,
-        description: 'Regional domain name for DNS CNAME/alias target',
-      });
-
-      new cdk.CfnOutput(this, 'ApiCustomDomainName', {
-        value: apiDomainName,
-        description: 'Custom domain name for the platform REST API',
-      });
-
-      new cdk.CfnOutput(this, 'ApiRegionalDomainName', {
-        value: apiCustomDomain.domainNameAliasDomainName,
-        description: 'Regional domain name — point DNS here',
-      });
-    }
-
-    const agentCoreGatewayRole = new iam.Role(this, 'AgentCoreGatewayExecutionRole', {
-      assumedBy: new iam.ServicePrincipal('bedrock-agentcore.amazonaws.com'),
-      description: 'Execution role for AgentCore Gateway interceptors',
-    });
-
-    const gatewayPolicyEngine = new cdk.CfnResource(this, 'AgentCoreGatewayPolicyEngine', {
-      type: 'AWS::BedrockAgentCore::PolicyEngine',
-      properties: {
-        Name: gatewayPolicyConfiguration.policyEngineName,
-        Description: 'Cedar policy engine for AgentCore Gateway tool authorization',
-        Tags: [
-          {
-            Key: 'stack',
-            Value: this.stackName,
-          },
-          {
-            Key: 'component',
-            Value: 'platform-gateway-policy',
-          },
-        ],
-      },
-    });
-
-    const gatewayDefaultPolicy = new cdk.CfnResource(this, 'AgentCoreGatewayDefaultPolicy', {
-      type: 'AWS::BedrockAgentCore::Policy',
-      properties: {
-        Name: gatewayPolicyConfiguration.policyName,
-        Description: 'Baseline Cedar policy for AgentCore Gateway',
-        PolicyEngineId: gatewayPolicyEngine.getAtt('PolicyEngineId').toString(),
-        ValidationMode: 'FAIL_ON_ANY_FINDINGS',
-        Definition: {
-          Cedar: {
-            Statement: [
-              'permit (',
-              '  principal,',
-              '  action,',
-              '  resource',
-              ') when {',
-              '  true',
-              '};',
-            ].join('\n'),
-          },
-        },
-      },
-    });
-    gatewayDefaultPolicy.addDependency(gatewayPolicyEngine);
-
-    agentCoreGatewayRole.addToPolicy(
-      new iam.PolicyStatement({
-        actions: ['lambda:InvokeFunction'],
-        resources: [this.requestInterceptorFn.functionArn, this.responseInterceptorFn.functionArn],
-      }),
-    );
-    agentCoreGatewayRole.addToPolicy(
-      new iam.PolicyStatement({
-        actions: [
-          'bedrock-agentcore:AuthorizeAction',
-          'bedrock-agentcore:PartiallyAuthorizeActions',
-          'bedrock-agentcore:GetPolicyEngine',
-        ],
-        resources: [gatewayPolicyEngine.ref],
-      }),
-    );
-
-    const agentCoreGateway = new cdk.CfnResource(this, 'AgentCoreGateway', {
-      type: 'AWS::BedrockAgentCore::Gateway',
-      properties: {
-        Name: `${this.stackName.toLowerCase().replace(/[^a-z0-9-]/g, '-')}-gateway`,
-        Description: 'Platform AgentCore Gateway with request/response interceptors',
-        AuthorizerType: 'AWS_IAM',
-        ProtocolType: 'MCP',
-        RoleArn: agentCoreGatewayRole.roleArn,
-        PolicyEngineConfiguration: {
-          Arn: gatewayPolicyEngine.ref,
-          Mode: gatewayPolicyConfiguration.enforcementMode,
-        },
-        InterceptorConfigurations: [
-          {
-            InterceptionPoints: ['REQUEST'],
-            InputConfiguration: {
-              PassRequestHeaders: true,
-            },
-            Interceptor: {
-              Lambda: {
-                Arn: this.requestInterceptorFn.functionArn,
-              },
-            },
-          },
-          {
-            InterceptionPoints: ['RESPONSE'],
-            InputConfiguration: {
-              PassRequestHeaders: true,
-            },
-            Interceptor: {
-              Lambda: {
-                Arn: this.responseInterceptorFn.functionArn,
-              },
-            },
-          },
-        ],
-        Tags: {
-          stack: this.stackName,
-          component: 'platform-gateway',
-        },
-      },
-    });
-    agentCoreGateway.addDependency(gatewayDefaultPolicy);
 
     new cdk.CfnOutput(this, 'AgentCoreGatewayPolicyMode', {
-      value: gatewayPolicyConfiguration.enforcementMode,
+      value: platformGateway.enforcementMode,
       description: 'Policy enforcement mode for the AgentCore Gateway',
     });
   }
