+++ infra/cdk/lib/platform-waf.ts
@@ -1,0 +1,98 @@
+import * as cdk from 'aws-cdk-lib';
+import * as apigateway from 'aws-cdk-lib/aws-apigateway';
+import * as wafv2 from 'aws-cdk-lib/aws-wafv2';
+import { Construct } from 'constructs';
+
+export interface PlatformWafProps {
+  readonly api: apigateway.RestApi;
+}
+
+export class PlatformWaf extends Construct {
+  public readonly apiWebAcl: wafv2.CfnWebACL;
+
+  constructor(scope: Construct, id: string, props: PlatformWafProps) {
+    super(scope, id);
+
+    this.apiWebAcl = new wafv2.CfnWebACL(this, 'ApiWebAcl', {
+      name: `${cdk.Stack.of(this).stackName}-api-waf`,
+      defaultAction: { allow: {} },
+      scope: 'REGIONAL',
+      visibilityConfig: {
+        cloudWatchMetricsEnabled: true,
+        metricName: `${cdk.Stack.of(this).stackName}-api-waf`,
+        sampledRequestsEnabled: true,
+      },
+      rules: [
+        {
+          name: 'AWSManagedRulesCommonRuleSet',
+          priority: 0,
+          overrideAction: { none: {} },
+          statement: {
+            managedRuleGroupStatement: {
+              vendorName: 'AWS',
+              name: 'AWSManagedRulesCommonRuleSet',
+            },
+          },
+          visibilityConfig: {
+            cloudWatchMetricsEnabled: true,
+            metricName: 'aws-managed-common',
+            sampledRequestsEnabled: true,
+          },
+        },
+        {
+          name: 'UkIpRateLimit',
+          priority: 1,
+          action: { block: {} },
+          statement: {
+            rateBasedStatement: {
+              aggregateKeyType: 'IP',
+              limit: 2000,
+              scopeDownStatement: {
+                geoMatchStatement: {
+                  countryCodes: ['GB'],
+                },
+              },
+            },
+          },
+          visibilityConfig: {
+            cloudWatchMetricsEnabled: true,
+            metricName: 'uk-ip-rate-limit',
+            sampledRequestsEnabled: true,
+          },
+        },
+        {
+          name: 'BlockSqlmapUserAgent',
+          priority: 2,
+          action: { block: {} },
+          statement: {
+            byteMatchStatement: {
+              fieldToMatch: {
+                singleHeader: {
+                  Name: 'user-agent',
+                },
+              },
+              positionalConstraint: 'CONTAINS',
+              searchString: 'sqlmap',
+              textTransformations: [
+                {
+                  priority: 0,
+                  type: 'LOWERCASE',
+                },
+              ],
+            },
+          },
+          visibilityConfig: {
+            cloudWatchMetricsEnabled: true,
+            metricName: 'block-sqlmap-user-agent',
+            sampledRequestsEnabled: true,
+          },
+        },
+      ],
+    });
+
+    new wafv2.CfnWebACLAssociation(this, 'ApiWebAclAssociation', {
+      resourceArn: props.api.deploymentStage.stageArn,
+      webAclArn: this.apiWebAcl.attrArn,
+    });
+  }
+}
